+++ README.md
@@ -23,7 +23,7 @@ Runnable with [`scala-cli`](https://scala-cli.virtuslab.org/).
 
 ```scala
 //> using scala "3.3.6"
-//> using dep "io.github.marcinzh::turbolift-core:0.118.0"
+//> using dep "io.github.marcinzh::turbolift-core:0.122.0"
 import turbolift.!!
 import turbolift.effects.{Reader, State, Error}
 
@@ -55,8 +55,8 @@ Same, but with [bindless](modules/bindless) syntax extension (`async/await` alik
 
 ```scala
 //> using scala "3.3.6"
-//> using dep "io.github.marcinzh::turbolift-core:0.118.0"
-//> using dep "io.github.marcinzh::turbolift-bindless:0.118.0"
+//> using dep "io.github.marcinzh::turbolift-core:0.122.0"
+//> using dep "io.github.marcinzh::turbolift-bindless:0.122.0"
 import turbolift.!!
 import turbolift.effects.{Reader, State, Error}
 import turbolift.bindless._
@@ -91,7 +91,7 @@ Same as the first example, but using monomorphized effect instances:
 
 ```scala
 //> using scala "3.3.6"
-//> using dep "io.github.marcinzh::turbolift-core:0.118.0"
+//> using dep "io.github.marcinzh::turbolift-core:0.122.0"
 import turbolift.!!
 import turbolift.effects.{ReaderEffect, StateEffect, ErrorEffect}
 
@@ -133,11 +133,11 @@ sbt examples/run
 ## Usage in SBT
 
 ```scala
-libraryDependencies += "io.github.marcinzh" %% "turbolift-core" % "0.118.0"
+libraryDependencies += "io.github.marcinzh" %% "turbolift-core" % "0.122.0"
 ```
 
 Optionally, for the [bindless](modules/bindless) syntax extension:
 
 ```scala
-libraryDependencies += "io.github.marcinzh" %% "turbolift-bindless" % "0.118.0"
+libraryDependencies += "io.github.marcinzh" %% "turbolift-bindless" % "0.122.0"
 ```
