+++ modules/core/src/main/scala/turbolift/internals/engine/Misc.scala
@@ -1,6 +1,8 @@
 package turbolift.internals.engine
 import turbolift.!!
 import turbolift.data.{Outcome, Exceptions}
+import turbolift.io.{Fiber, Zipper, Warp}
+import turbolift.interpreter.Continuation
 
 
 //@#@
@@ -15,3 +17,17 @@ private[engine] object Misc:
   def impossible: Nothing = panic("impossible happened")
 
   def nullableToOption[T](x: T | Null): Option[T] = if x != null then Some(x) else None
+
+
+
+  extension (thiz: Continuation[?, ?, ?, ?])
+    private[engine] inline def asImpl: ContImpl = thiz.asInstanceOf[ContImpl]
+
+  extension (thiz: Fiber[?, ?])
+    private[engine] inline def asImpl: FiberImpl = thiz.asInstanceOf[FiberImpl]
+
+  extension (thiz: Zipper[?, ?])
+    private[engine] inline def asImpl: ZipperImpl = thiz.asInstanceOf[ZipperImpl]
+
+  extension (thiz: Warp)
+    private[engine] inline def asImpl: WarpImpl = thiz.asInstanceOf[WarpImpl]
